requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1

# Vector Store
chromadb>=0.4.22

# Embeddings and LLM
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0

# Document Loading
pypdf>=3.17.0
python-magic-bin>=0.4.14  # For Windows
pymupdf>=1.23.0

# Configuration
pyyaml>=6.0.1

# CLI
click>=8.1.7

# Evaluation (optional - for advanced metrics)
# ragas>=0.1.0
# datasets>=2.16.0

# Additional utilities
tqdm>=4.66.0
numpy>=1.24.0
colorama>=0.4.6  # For colored terminal output
